Move MouseTrackerAnnotation to services
