+++ packages/flutter/lib/src/rendering/proxy_box.dart
@@ -15,7 +15,7 @@ import 'package:vector_math/vector_math_64.dart';
 import 'box.dart';
 import 'layer.dart';
 import 'layout_helper.dart';
-import 'mouse_tracking.dart';
+import 'mouse_tracker.dart';
 import 'object.dart';
 
 export 'package:flutter/gestures.dart' show
